semi final

+++ Core/Hotel Management.ServiceImplementiton/Services/HotelService/HotelFeatures.cs
@@ -6,11 +6,11 @@ using Hotel_Management.DOMAIN.Models.HotelModel;
 using Hotel_Management.ServiceAbstraction.IserviceOfHotel;
 using Hotel_Management.Shared.DTOs.Hotel.Features;
 using Hotel_Management.Shared.DTOs.Hotel.HotelFeaturesDtos;
-
-
+using Microsoft.AspNetCore.Http;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -21,19 +21,36 @@ namespace Hotel_Management.ServiceImplementiton.Services.HotelService
         private readonly IUow uow;
         private readonly IHotelFeatureReposatory hotelrepo;
         private readonly IMapper mapper;
-        
+        private readonly IHttpContextAccessor httpContextAccessor;
 
-        public HotelFeatureservice(IUow uow, IHotelFeatureReposatory hotelrepo,IMapper mapper)
+        public HotelFeatureservice(IUow uow, IHotelFeatureReposatory hotelrepo,IMapper mapper,IHttpContextAccessor _httpContextAccessor)
         {
             this.uow = uow;
             this.hotelrepo = hotelrepo;
             this.mapper = mapper;
-           
+            httpContextAccessor = _httpContextAccessor;
         }
 
         public async Task<int> Add(AddHotelFeaturesDto dto)
         {
-            await hotelrepo.AddHotelFeaturesAsync(dto.FeatureIds,dto.HotelId);
+            var Hotel = await uow.GenerateRepo<Hotel, int>().GetById(dto.HotelId);
+            if(Hotel is not null)
+            {
+                if (GetuserId() == Hotel.managerId)
+                {
+                    await hotelrepo.AddHotelFeaturesAsync(dto.FeatureIds, dto.HotelId);
+
+                }
+                else
+                {
+                    throw new Exception("not allowed to you");
+                }
+            }
+            else
+            {
+                throw new Exception("Hotel is not found");
+
+            }
             return await uow.SaveChanges();
         }
 
@@ -57,9 +74,56 @@ namespace Hotel_Management.ServiceImplementiton.Services.HotelService
 
         public async Task<int> RemoveFeatureFromHotelAsync(int hotelId, int featureId)
         {
-          await hotelrepo.Delete(featureId, hotelId);
+            var Hotel = await uow.GenerateRepo<Hotel, int>().GetById(hotelId);
+
+            if (Hotel is not null)
+            {
+                if (GetuserId() == Hotel.managerId|| GetuserRole()=="SuperAdmin")
+                {
+                    await hotelrepo.Delete(featureId, hotelId);
+
+                }
+                else
+                {
+                    throw new Exception("not allowed to you");
+                }
+            }
+            else
+            {
+                throw new Exception("Hotel is not found");
+
+            }
             return await uow.SaveChanges();
 
         }
+        private Guid GetuserId()
+        {
+            var user = httpContextAccessor.HttpContext?.User;
+
+            if (user == null || !user.Identity.IsAuthenticated)
+                throw new UnauthorizedAccessException("User is not authenticated");
+
+            var userIdClaim = user.FindFirst(ClaimTypes.NameIdentifier);
+
+            if (userIdClaim == null)
+                throw new UnauthorizedAccessException("UserId not found in token");
+
+            return Guid.Parse(userIdClaim.Value);
+        }
+        private string GetuserRole()
+        {
+            var user = httpContextAccessor.HttpContext?.User;
+
+            if (user == null || !user.Identity.IsAuthenticated)
+                throw new UnauthorizedAccessException("User is not authenticated");
+
+            var userroleClaim = user.FindFirst(ClaimTypes.Role);
+
+            if (userroleClaim == null)
+                throw new UnauthorizedAccessException("UserId not found in token");
+
+            return userroleClaim.Value;
+        }
+        
     }
 }
